pipeline/pyproject.toml
[project]
name = "izakaya-pipeline"
version = "0.1.0"
description = "Izakaya Dagster pipeline"
requires-python = ">=3.12"
dependencies = [
    "dagster>=1.9",
    "dagster-cloud",
    "dagster-dbt>=0.24",
    "dbt-bigquery>=1.8",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "google-cloud-bigquery>=3.25",
    "pandas>=2.0",
    "db-dtypes>=1.0",
    "anthropic>=0.50",
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.dagster]
module_name = "izakaya_pipeline.definitions"
